+++ clang-tidy/misc/MiscTidyModule.cpp
@@ -10,6 +10,8 @@
 #include "../ClangTidy.h"
 #include "../ClangTidyModule.h"
 #include "../ClangTidyModuleRegistry.h"
+#include "ClassInheritFromStructCheck.h"
+#include "ClassMemberDataAccessCheckCheck.h"
 #include "DefinitionsInHeadersCheck.h"
 #include "FunctionsInStructCheck.h"
 #include "MisplacedConstCheck.h"
@@ -31,6 +33,11 @@ namespace misc {
 class MiscModule : public ClangTidyModule {
 public:
   void addCheckFactories(ClangTidyCheckFactories &CheckFactories) override {
+
+    CheckFactories.registerCheck<ClassInheritFromStructCheck>(
+        "misc-class-inherit-from-struct");
+    CheckFactories.registerCheck<ClassMemberDataAccessCheckCheck>(
+        "misc-class-member-data-access-check");
     CheckFactories.registerCheck<FunctionsInStructCheck>(
         "misc-functions-in-struct");
     CheckFactories.registerCheck<MisplacedConstCheck>("misc-misplaced-const");