+++ docs/ReleaseNotes.rst
@@ -57,8 +57,13 @@ The improvements are...
 Improvements to clang-tidy
 --------------------------
 
+- New :doc:`misc-class-inherit-from-struct
+  <clang-tidy/checks/misc-class-inherit-from-struct>` check.
+- New :doc:`misc-class-member-data-access-check
+  <clang-tidy/checks/misc-class-member-data-access-check>` check.
 - New :doc:`misc-functions-in-struct
   <clang-tidy/checks/misc-functions-in-struct>` check.
+
 
   FIXME: add release notes.